Update wording, add min-height
