+++ packages/bodiless-organisms/src/components/FilterByGroup/Filter.tsx
@@ -152,7 +152,7 @@ const TagTitle = flow(
   withTagProps({
     allowMultipleTags: false,
     placeholder: 'Add or Create',
-    formTitle: 'Groups',
+    formTitle: 'Group Membership',
     formBodyText: 'Select from available groups:',
     seeAllText: 'View All Groups',
     noSuggestionsText: 'No matching groups found.',
